requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.10.2
anyio==3.6.2
apprise==1.3.0
asgi-lifespan==2.1.0
astroid==2.15.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
cachetools==5.3.0
camelot-py==0.11.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.3.8
cryptography==40.0.1
dateparser==1.1.8
dill==0.3.6
docker==6.0.1
et-xmlfile==1.1.0
fastapi==0.95.0
frozenlist==1.3.3
fsspec==2023.1.0
fuzzywuzzy==0.18.0
ghostscript==0.7
google-auth==2.17.0
greenlet==2.0.2
griffe==0.25.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kubernetes==26.1.0
lazy-object-proxy==1.9.0
Levenshtein==0.20.9
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
orjson==3.8.9
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
pdfminer.six==20221105
pendulum==2.1.2
platformdirs==3.2.0
pluggy==1.0.0
prefect==2.8.7
prefect-slack==0.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pylint==2.17.1
PyMuPDF==1.21.1
pyodbc==4.0.35
pypdf==3.7.0
PyPDF2==3.0.1
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.20.9
python-slugify==8.0.1
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
pywin32==306
PyYAML==6.0
rapidfuzz==2.15.0
readchar==4.0.5
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.3
rsa==4.9
six==1.16.0
slack-sdk==3.20.2
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.26.1
tabulate==0.9.0
text-unidecode==1.3
tika==2.6.0
toml==0.10.2
tomlkit==0.11.7
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
websocket-client==1.5.1
websockets==10.4
wrapt==1.15.0
yarl==1.8.2
